Split TestDateTimeOperators into variants based on isLegacyTimestamp
